refactor: 重组江山如故系列武将包
