+++ lang/zh_CN/Package/CountryAsBeforeElucidationPackage.lua
@@ -1,0 +1,14 @@
+-- translation for Country As Before Elucidation Package
+
+return {
+    ['CountryAsBeforeElucidationPackage'] = '江山如故·承',
+    ['ExChengGuanyu'] = '关羽',
+    ['&ExChengGuanyu'] = '关羽',
+    ['#ExChengGuanyu'] = '羊左之义',
+    ['LuaGuanjue'] = '冠绝',
+    [':LuaGuanjue'] = '锁定技，当你使用或打出一张牌时，你令所有其他角色于此回合内不能使用或打出花色与之相同的牌',
+    ['LuaNianen'] = '念恩',
+    [':LuaNianen'] = '你可以将你的一张牌当任意基本牌使用或打出，然后若你以此法转化后的牌不为普通【杀】或此牌不为红色，则直到此回合结束，你视为拥有“马术”且不能发动本技能',
+    ['nianen_slash'] = '念恩',
+    ['nianen_saveself'] = '念恩',
+}
