+++ lang/zh_CN/Package/ExpansionPackage.lua
@@ -568,15 +568,6 @@ return {
     [':LuaGuli'] = '出牌阶段限一次，你可以将所有手牌当一张【杀】使用（此【杀】无视目标角色的防具）。然后此【杀】结算结束后，若此【杀】造成过伤害，你可以失去1点体力，将手牌摸至X张（X为你的体力上限）',
     ['LuaAosi'] = '骜肆',
     [':LuaAosi'] = '锁定技，当你于出牌阶段内对其他角色造成伤害后，若其在你攻击范围内，你本阶段对其使用牌无次数限制',
-    ['ExChengGuanyu'] = '承·关羽',
-    ['&ExChengGuanyu'] = '承关羽',
-    ['#ExChengGuanyu'] = '羊左之义',
-    ['LuaGuanjue'] = '冠绝',
-    [':LuaGuanjue'] = '锁定技，当你使用或打出一张牌时，你令所有其他角色于此回合内不能使用或打出花色与之相同的牌',
-    ['LuaNianen'] = '念恩',
-    [':LuaNianen'] = '你可以将你的一张牌当任意基本牌使用或打出，然后若你以此法转化后的牌不为普通【杀】或此牌不为红色，则直到此回合结束，你视为拥有“马术”且不能发动本技能',
-    ['nianen_slash'] = '念恩',
-    ['nianen_saveself'] = '念恩',
     ['ExZhugezhan'] = '诸葛瞻',
     ['&ExZhugezhan'] = '诸葛瞻',
     ['#ExZhugezhan'] = '临难死义',
